Added pre-commit hook + update Makefile
The pre-commit hook is added to ensure the correct HTML
file is published. The reason for that is this: sometimes
I do `git add --patch` and the some of the changes are not
staged. I want the final HTML file be compiled based on
the one I staged not the one in the working directory.
Thus with `make publish`, I make sure I'm compiling the
correct `README.md`.

Makefile updates:
- `setuphook` target in Makefile
- Re-structure targets; now more used targets are near
  the top of the file
- Addition of `make publish` to build the HTML based on
  the staged file
  - To get the file in the index
    src: https://stackoverflow.com/a/60854287
    ```
    git show :0:path/to/file > staged_file
    #or
    git show :path/to/some/file
    ```